Update `core`
`implement_model` move to  _wrapper.py
